requirements.txt
# FastAPI et serveur
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
pydantic==2.5.3

# Traitement audio
numpy==1.24.3
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
torchcodec>=0.0.1

# Reconnaissance du langage et transcription
openai-whisper

# Diarisation
pyannote.audio==4.0.4
pyannote.core==6.0.1

# Machine Learning
scikit-learn==1.3.2
transformers>=4.30.0
datasets>=2.14.0
evaluate>=0.4.0
accelerate>=0.20.0

# Visualisation (pour évaluation)
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilitaires
python-dotenv==1.2.1

# Data processing
pandas==3.0.1
tqdm==4.67.3
